Allow RubyGems to link to homepage & source code
Currently it's insanely difficult to find the source code repository for 
jsonapi_suite on [rubygems.org][1]. 

I'd recommend publishing a 0.7.1 patch release just with this gemspec 
change so people can find the source repository and homepage and 
potentially fork jsonapi_suite to allow for Rails 6 if they can't move to 
graphiti right away for some reason. 

I'm guessing it may not be possible to loosen the version requirement 
on actionpack and activesupport but hopefully it is. 

[1]: https://rubygems.org/gems/jsonapi_suite/versions/0.7.0